[HELPS-981] Refactor type specific behaviour
